Cleanup of tool
