+++ packages/pwa-kit-mcp/src/server/server.js
@@ -14,8 +14,7 @@ import {
     CreateNewComponentTool,
     DeveloperGuidelinesTool,
     TestWithPlaywrightTool,
-    CreateNewPageTool,
-    VersionControlGitTool
+    CreateNewPageTool
 } from '../tools'
 
 // NOTE: This is a workaround to import JSON files as ES modules.
@@ -39,7 +38,7 @@ class PwaStorefrontMCPServerHighLevel {
             }
         )
         this.createNewComponentTool = new CreateNewComponentTool()
-        this.versionControlGitTool = new VersionControlGitTool()
+        this.createAppGuidelinesTool = new CreateAppGuidelinesTool()
         this.testWithPlaywrightTool = new TestWithPlaywrightTool()
         this.setupTools()
     }
@@ -47,10 +46,10 @@ class PwaStorefrontMCPServerHighLevel {
     setupTools() {
         // Register CreateProjectTool
         this.server.tool(
-            CreateAppGuidelinesTool.name,
-            CreateAppGuidelinesTool.description,
-            CreateAppGuidelinesTool.inputSchema,
-            CreateAppGuidelinesTool.fn
+            this.createAppGuidelinesTool.name,
+            this.createAppGuidelinesTool.description,
+            this.createAppGuidelinesTool.inputSchema,
+            this.createAppGuidelinesTool.fn
         )
         this.server.tool(
             DeveloperGuidelinesTool.name,
@@ -78,12 +77,6 @@ class PwaStorefrontMCPServerHighLevel {
             CreateNewPageTool.description,
             CreateNewPageTool.inputSchema,
             CreateNewPageTool.handler
-        )
-        this.server.tool(
-            this.versionControlGitTool.name,
-            this.versionControlGitTool.description,
-            this.versionControlGitTool.inputSchema,
-            this.versionControlGitTool.handler
         )
     }
